python/bican_mccarroll_helloworld/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
[project]
name = "bican_mccarroll_helloworld"
version = "0.1.0"
dependencies = [
    "Biopython",
]
authors = [
  { name="Alec Wysoker", email="alecw@broadinstitute.org" },
]
description = "Example python package."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/broadinstitute/bican-mccarroll-manuscript1/"
Issues = "https://github.com/broadinstitute/bican-mccarroll-manuscript1/issues"

[project.scripts]
bican_mccarroll_hello = "bican_mccarroll_helloworld.hello:main"
